some clang support

+++ appframework/glmdisplaydb_linuxwin.inl
@@ -9,6 +9,8 @@
 
 //===============================================================================
 
+#include "stdint.h"
+
 GLMRendererInfo::GLMRendererInfo( void )
 {
         m_display = NULL;
@@ -621,7 +623,7 @@ void    GLMDisplayInfo::Dump( int which )
         SDLAPP_FUNC;
 
         GLMPRINTF(("\n         #%d: GLMDisplayInfo @ %08x, pixwidth=%d  pixheight=%d",
-                           which, (int)this,  m_info.m_displayPixelWidth,  m_info.m_displayPixelHeight ));
+                           which, (uintptr_t)this,  m_info.m_displayPixelWidth,  m_info.m_displayPixelHeight ));
 
         FOR_EACH_VEC( *m_modes, i )
         {
